refactor sanity plot code
* move SanityPlotWidget to new file
